requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
dataclasses-json==0.6.4
deepgram-sdk==3.2.3
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.16.2
packaging==24.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
uvicorn==0.29.0
verboselogs==1.7
websockets==12.0
yarl==1.9.4
